+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 5.6.2
 ----------------
 - Fix SQL error on delete when current Student/Staff ID was lost (in other browser tab) in Student.php & User.php
+- Fix SQL error foreign keys: check if can DELETE from schools and staff in Rollover.php
 
 Changes in 5.6.1
 ----------------